throwing in appendLeaves + tests
