+++ yarn-project/merkle-tree/src/test/standard_based_test_suite.ts
@@ -1,0 +1,117 @@
+import { BarretenbergWasm } from '@aztec/barretenberg.js/wasm';
+import { default as levelup } from 'levelup';
+import { Hasher, Pedersen, SiblingPath, StandardMerkleTree } from '../index.js';
+import { SparseMerkleTree } from '../sparse_tree/sparse_tree.js';
+import { appendLeaves, createMemDown } from './utils.js';
+
+export const standardBasedTreeTestSuite = (
+  testName: string,
+  createDb: (
+    levelup: levelup.LevelUp,
+    hasher: Hasher,
+    name: string,
+    depth: number,
+  ) => Promise<StandardMerkleTree | SparseMerkleTree>,
+  appendImplemented = true,
+) => {
+  describe(testName, () => {
+    let wasm: BarretenbergWasm;
+    let pedersen: Pedersen;
+    const values: Buffer[] = [];
+
+    beforeAll(() => {
+      for (let i = 0; i < 4; ++i) {
+        const v = Buffer.alloc(32, i + 1);
+        v.writeUInt32BE(i, 28);
+        values[i] = v;
+      }
+    });
+    beforeEach(async () => {
+      wasm = await BarretenbergWasm.get();
+      pedersen = new Pedersen(wasm);
+    });
+
+    it('should have correct empty tree root for depth 32', async () => {
+      const db = levelup(createMemDown());
+      const tree = await createDb(db, pedersen, 'test', 32);
+      const root = tree.getRoot();
+      expect(root.toString('hex')).toEqual('18ceb5cd201e1cee669a5c3ad96d3c4e933a365b37046fc3178264bede32c68d');
+    });
+
+    it('should have correct root and sibling paths', async () => {
+      const db = levelup(createMemDown());
+      const tree = await createDb(db, pedersen, 'test', 2);
+
+      const zeroTreeLeafHash = StandardMerkleTree.ZERO_ELEMENT;
+      const level1ZeroHash = pedersen.compress(zeroTreeLeafHash, zeroTreeLeafHash);
+      expect(tree.getNumLeaves()).toEqual(0n);
+      expect(tree.getRoot()).toEqual(pedersen.compress(level1ZeroHash, level1ZeroHash));
+      expect(await tree.getSiblingPath(0n)).toEqual(new SiblingPath([zeroTreeLeafHash, level1ZeroHash]));
+
+      await appendLeaves(appendImplemented, tree, [values[0]]);
+      expect(tree.getNumLeaves(true)).toEqual(1n);
+      expect(tree.getNumLeaves(false)).toEqual(0n);
+      expect(tree.getRoot(true)).toEqual(
+        pedersen.compress(pedersen.compress(values[0], zeroTreeLeafHash), level1ZeroHash),
+      );
+      expect(await tree.getSiblingPath(0n, true)).toEqual(new SiblingPath([zeroTreeLeafHash, level1ZeroHash]));
+      expect(tree.getRoot(false)).toEqual(pedersen.compress(level1ZeroHash, level1ZeroHash));
+      expect(await tree.getSiblingPath(0n, false)).toEqual(new SiblingPath([zeroTreeLeafHash, level1ZeroHash]));
+
+      await appendLeaves(appendImplemented, tree, [values[1]]);
+      expect(tree.getNumLeaves(true)).toEqual(2n);
+      expect(tree.getRoot(true)).toEqual(pedersen.compress(pedersen.compress(values[0], values[1]), level1ZeroHash));
+      expect(await tree.getSiblingPath(1n, true)).toEqual(new SiblingPath([values[0], level1ZeroHash]));
+      expect(tree.getNumLeaves(false)).toEqual(0n);
+      expect(tree.getRoot(false)).toEqual(pedersen.compress(level1ZeroHash, level1ZeroHash));
+      expect(await tree.getSiblingPath(1n, false)).toEqual(new SiblingPath([zeroTreeLeafHash, level1ZeroHash]));
+
+      await appendLeaves(appendImplemented, tree, [values[2]]);
+      expect(tree.getNumLeaves(true)).toEqual(3n);
+      expect(tree.getRoot(true)).toEqual(
+        pedersen.compress(pedersen.compress(values[0], values[1]), pedersen.compress(values[2], zeroTreeLeafHash)),
+      );
+      expect(await tree.getSiblingPath(2n, true)).toEqual(
+        new SiblingPath([zeroTreeLeafHash, pedersen.compress(values[0], values[1])]),
+      );
+      expect(tree.getNumLeaves()).toEqual(0n);
+      expect(tree.getRoot()).toEqual(pedersen.compress(level1ZeroHash, level1ZeroHash));
+      expect(await tree.getSiblingPath(2n)).toEqual(new SiblingPath([zeroTreeLeafHash, level1ZeroHash]));
+
+      await appendLeaves(appendImplemented, tree, [values[3]]);
+      expect(tree.getNumLeaves(true)).toEqual(4n);
+      expect(tree.getRoot(true)).toEqual(
+        pedersen.compress(pedersen.compress(values[0], values[1]), pedersen.compress(values[2], values[3])),
+      );
+      expect(await tree.getSiblingPath(3n, true)).toEqual(
+        new SiblingPath([values[2], pedersen.compress(values[0], values[1])]),
+      );
+      expect(tree.getNumLeaves(false)).toEqual(0n);
+      expect(tree.getRoot(false)).toEqual(pedersen.compress(level1ZeroHash, level1ZeroHash));
+      expect(await tree.getSiblingPath(3n, false)).toEqual(new SiblingPath([zeroTreeLeafHash, level1ZeroHash]));
+      // Lifted from memory_tree.test.cpp to ensure consistency.
+      //expect(root.toString('hex')).toEqual('0bf2e78afd70f72b0e6eafb03c41faef167a82441b05e517cdf35d813302061f');
+      expect(await tree.getSiblingPath(0n, true)).toEqual(
+        new SiblingPath([values[1], pedersen.compress(values[2], values[3])]),
+      );
+      expect(await tree.getSiblingPath(1n, true)).toEqual(
+        new SiblingPath([values[0], pedersen.compress(values[2], values[3])]),
+      );
+      expect(await tree.getSiblingPath(2n, true)).toEqual(
+        new SiblingPath([values[3], pedersen.compress(values[0], values[1])]),
+      );
+      expect(await tree.getSiblingPath(3n, true)).toEqual(
+        new SiblingPath([values[2], pedersen.compress(values[0], values[1])]),
+      );
+
+      await tree.commit();
+      // now committed state should equal uncommitted state
+      expect(await tree.getSiblingPath(0n, false)).toEqual(await tree.getSiblingPath(0n, true));
+      expect(await tree.getSiblingPath(1n, false)).toEqual(await tree.getSiblingPath(1n, true));
+      expect(await tree.getSiblingPath(2n, false)).toEqual(await tree.getSiblingPath(2n, true));
+      expect(await tree.getSiblingPath(3n, false)).toEqual(await tree.getSiblingPath(3n, true));
+      expect(tree.getNumLeaves(false)).toEqual(tree.getNumLeaves(true));
+      expect(tree.getRoot(false)).toEqual(tree.getRoot(true));
+    });
+  });
+};
